+++ src/bgame/bg_public.h
@@ -542,7 +542,7 @@ enum MissileFlightMode : __int32
     MISSILEFLIGHTMODE_DIRECT = 0x1,
 };
 
-enum team_t;
+enum team_t : __int32; // ISO C++ requires underlying type on forward enum decls
 #ifdef KISAK_MP
 struct corpse_ent_t // sizeof=0x4
 {                                       // ...